requirements.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
packaging==23.2
pkgutil_resolve_name==1.3.10
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
rpds-py==0.13.2
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
zipp==3.17.0
